Rearranged include configuration

+++ Src/Shape/Rectangle.h
@@ -1,3 +1,5 @@
+#include "..\Main.h"
+
 #ifndef _aitest_RECTANGLE_H
 #define _aitest_RECTANGLE_H
 
@@ -6,7 +8,9 @@ class Rectangle : public Shape
 public:
 	Rectangle(Vector2D<double> pos, double width, double height)
 	{
-		mShape = sf::Shape::Rectangle(pos.x, pos.y, pos.x+width, pos.y+height, sf::Color(0,0,0,255));
+		mShape = sf::Shape::Rectangle((float)pos.x, (float)pos.y,
+										(float)(pos.x+width), (float)(pos.y+height),
+										sf::Color(0,0,0,255));
 		SetPos(pos);
 	}
